image_generator/requirements_long.txt
absl-py==1.0.0
alembic==1.7.6
antlr4-python3-runtime==4.8
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
cycler==0.11.0
databricks-cli==0.16.4
decorator==4.4.2
dill==0.3.4
docker==5.0.3
dominate==2.6.0
entrypoints==0.4
Flask==2.0.3
fonttools==4.29.1
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.0
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.44.0
gunicorn==20.1.0
idna==3.3
imageio==2.16.1
imageio-ffmpeg==0.4.5
importlib-metadata==4.11.2
importlib-resources==5.4.0
itsdangerous==2.1.0
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
lpips==0.1.4
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
mlflow==1.24.0
monai==0.8.1
moviepy==1.0.3
networkx==2.7
nibabel==3.2.2
numpy==1.22.2
oauthlib==3.2.0
omegaconf==2.1.1
packaging==21.3
pandas==1.4.1
patsy==0.5.2
Pillow==9.0.1
proglog==0.1.9
prometheus-client==0.13.1
prometheus-flask-exporter==0.18.7
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
querystring-parser==1.2.4
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
SQLAlchemy==1.4.31
sqlparse==0.4.2
statsmodels==0.13.2
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
threadpoolctl==3.1.0
tifffile==2022.2.9
torch==1.10.2
torchvision==0.11.3
tqdm==4.63.0
typing_extensions==4.1.1
urllib3==1.26.8
websocket-client==1.3.1
Werkzeug==2.0.3
zipp==3.7.0
